fix: fix the logic to update the DB
Updates happen right after detecting a change. This is wrong because
once the call the CDN API to purge files (to be written) may fail and
then the correct thing is to NOT update the DB.

As an aside, updating in the thread that computes the checksum
effectively makes all those thread queue behind one another, because
writers are sequential in Sqlite. This patch fixes this as well.